refactor: change route name from product-catalog to products
