+++ src/app/products/[productSlug]/page.tsx
