+++ src/components/products/searchSection.tsx
